Version 3.0.0 Update
Major overhaul - new version. See arxiv.org/abs/1708.05305 for full
details

+++ elecsus/libs/ang_mom_p.py
@@ -1,4 +1,4 @@
-# Copyright 2014 M. A. Zentile, J. Keaveney, L. Weller, D. Whiting,
+# Copyright 2017 M. A. Zentile, J. Keaveney, L. Weller, D. Whiting,
 # C. S. Adams and I. G. Hughes.
 
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,25 +13,25 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""return J_+ matrix for angular momentum j
+"""
+return J_+ matrix for angular momentum j
 
 Calculates the + ladder operator for a given angular momentum.
 
 Called by ang_mom.py
-
 """
 
 from numpy import zeros,sqrt,arange
 
 def jp(jj):
-    b=0
-    dim=2.0*jj+1.0
-    jp=zeros((dim,dim))
-    z=arange(dim)
-    m=jj-z
+    b = 0
+    dim = int(2*jj+1)
+    jp = zeros((dim,dim))
+    z = arange(dim)
+    m = jj-z
     while b<dim-1.0:
-        mm=m[b+1]
-        jp[b,b+1]=sqrt(jj*(jj+1)-mm*(mm+1)) 
-        b=b+1.0
+        mm = m[b+1]
+        jp[b,b+1] = sqrt(jj*(jj+1)-mm*(mm+1)) 
+        b = b+1
     return jp
